bug 477706 - fixing most VC 2005 warnings. r=gal,mrbkap

+++ js/src/jsscript.cpp
@@ -508,8 +508,8 @@ js_XDRScript(JSXDRState *xdr, JSScript **scriptp, JSBool *hasMagic)
             return JS_FALSE;
 
         script->main += prologLength;
-        script->version = (JSVersion) (version & 0xffff);
-        script->nfixed = (uint16) (version >> 16);
+        script->version = JSVersion(version & 0xffff);
+        script->nfixed = uint16(version >> 16);
 
         /* If we know nsrcnotes, we allocated space for notes in script. */
         notes = SCRIPT_NOTES(script);
@@ -582,7 +582,7 @@ js_XDRScript(JSXDRState *xdr, JSScript **scriptp, JSBool *hasMagic)
         }
         script->lineno = (uintN)lineno;
         script->nslots = (uint16)nslots;
-        script->staticDepth = nslots >> 16;
+        script->staticDepth = (uint16)(nslots >> 16);
     }
 
     for (i = 0; i != natoms; ++i) {
